Refactor _calc_slice
